+++ src/main/java/ru/hogwarts/school/service/StudentService.java
@@ -10,11 +10,11 @@ public interface StudentService {
 //    Student addStudent(Student student);
     Student createStudent(Student student);
 
-    Student findStudent(long id);
+    Student findStudent(Long id);
 
     Student editStudent(Student student);
 
-    void deleteStudent(long id);
+    void deleteStudent(Long id);
 
-    public Collection<Student> findByAge(int age);
+    public Collection<Student> findByAge(Long age);
 }
